Implement verb loop and recipe invocation
All instruments used in chef_fib.txt have been implemented, though this
interpreter outputs wrong fibonacci numbers when running it.
Based on a comparison with the perl implementation (Acme::Chef), it
seems that this behavior is correct. Maybe original code has something
wrong? (And whats more, the calculation is impractical since the order
of recursive fibonacci is O(2^n). What a trap!)